Only normalize WSO21 fixture file

+++ spec/fixtures/files/sac_imports_src/export_sac_imports_fixture.sh
@@ -9,6 +9,10 @@ for file in sac_imports_fixture-*.csv; do
     sheet_name=$(echo "$file" | cut -d'-' -f2 | cut -d'.' -f1)
 
     # Normalize and rename the file to [sheet_name]_[original_file_name].csv
-    ./normalize_csv.rb "$file" > "${sheet_name}_fixture.csv"
-    rm "$file"
+    if [[ "$file" =~ WSO2 ]]; then
+        ./normalize_csv.rb "$file" > "${sheet_name}_fixture.csv"
+        rm "$file"
+    else
+        mv "$file" "${sheet_name}_fixture.csv"
+    fi
 done
